Fixed Main and upgraded to typescript some common files
