+++ .eslintrc.json
@@ -27,7 +27,8 @@
     "@typescript-eslint/no-var-requires": 0,
     "semi": "off",
     "@typescript-eslint/semi": "error",
-    "no-use-before-define": 0
+    "no-use-before-define": 0,
+    "@typescript-eslint/no-use-before-define": "off"
   },
   "settings": {
     "react": {
