+++ src/common/chat/submit-message/index.js
@@ -1,0 +1,3 @@
+import all from './submit-message.tsx';
+
+export default all;
