+++ src/common/rightBlock/index.js
@@ -1,15 +1,3 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-import { RightBlockContainer } from '../styledComponents';
-
-const RightBlock = (props) => {
-  const { children } = props;
-  return <RightBlockContainer>{children}</RightBlockContainer>;
-};
-
-RightBlock.propTypes = {
-  children: PropTypes.node.isRequired,
-};
+import RightBlock from './index.tsx';
 
 export default RightBlock;
