--- src/components/Main/ArticleForm/index.css
@@ -1,46 +1,0 @@
-.inputName {
-    margin-left: 0;
-    margin-top: 32px;
-    margin-bottom: 20px;
-    width: 1024px;
-    border: none;
-    border-bottom: 1px solid #B3B3B3;
-    outline: none !important;
-    font-size: 24px;
-}
-
-.inputText {
-    margin-left: 0;
-    margin-top: 42px;
-    width: 1024px;
-    border-radius: 5px;
-    border: 1px solid #B2B2B2;
-    outline: none !important;
-    font-size: 24px;
-    resize: none;
-}
-
-.articleForm {
-    display: flex;
-    flex-direction: column;
-    position: relative;
-}
-
-.articleButton {
-    width: 331px;
-    height: 67px;
-    margin-top: 100px;
-    background: #FFB11B;
-    border-radius: 15px;
-    border: none;
-    font-family: Montserrat,serif;
-    font-style: normal;
-    font-weight: 500;
-    font-size: 20px;
-    line-height: 29px;
-    letter-spacing: 0.065em;
-    color: #000000;
-}
-.articleButton:hover {
-    cursor: pointer;
-}
