+++ src/components/Main/Articles/blockComment/BlockComments/index.js
@@ -1,0 +1,3 @@
+import BlockComments from './BlockComments';
+
+export default BlockComments;
