+++ src/components/Main/Articles/blockComment/BlockComments/styles.js
@@ -1,0 +1,21 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+width: 100%;
+padding: 40px 0 40px 60px;
+display: flex;
+flex-direction: column;
+justify-content: flex-start;
+align-items: flex-start;
+z-index: 100;
+background: white;
+`;
+
+export const Title = styled.div`
+font-weight: 500;
+font-size: 20px;
+line-height: 160.9%;
+color: #515151;
+margin: 0 0 40px 0;
+font-family: Montserrat, serif;
+`;
