+++ src/components/Main/Articles/blockComment/ComponentInput/index.js
@@ -1,0 +1,3 @@
+import all from './ComponentInput.tsx';
+
+export default all;
