+++ src/components/Main/UserInfoHeader/index.js
@@ -1,32 +1,3 @@
-import React from 'react';
-
-import avatar from './img/main photo.png';
-import addPhotoIcon from './img/add photo.svg';
-import {
-  UserInfoHeaderContainer,
-  UserInfoAvatar,
-  UserInfoNameBlock,
-  Avatar,
-  AddPhotoIcon,
-  UserName,
-  UserProfession,
-  UserOnlineStatus,
-  UserOnlineIcon,
-} from '../../../common/styledComponents';
-
-const UserInfoHeader = () => (
-  <UserInfoHeaderContainer>
-    <UserInfoAvatar>
-      <Avatar img={avatar} />
-      <AddPhotoIcon img={addPhotoIcon} />
-      <UserOnlineIcon />
-    </UserInfoAvatar>
-    <UserInfoNameBlock>
-      <UserName>Богдан Леста</UserName>
-      <UserProfession>Frontend разработчик</UserProfession>
-      <UserOnlineStatus>был 20.07.2020 в 15:07</UserOnlineStatus>
-    </UserInfoNameBlock>
-  </UserInfoHeaderContainer>
-);
+import UserInfoHeader from './index.tsx';
 
 export default UserInfoHeader;
