+++ src/components/Main/Wall/FormStatus/index.js
@@ -1,32 +1,3 @@
-import React from 'react';
-import { useFormik } from 'formik';
+import comp from './index.tsx';
 
-import './index.css';
-
-const FormStatus = () => {
-  const formik = useFormik({
-    initialValues: {
-      statusText: '',
-    },
-    onSubmit: (values) => {
-      // eslint-disable-next-line no-console
-      console.log(values);
-    },
-  });
-
-  return (
-    <form className="statusForm" onSubmit={formik.handleSubmit}>
-      <input
-        className="statusText"
-        id="statusText"
-        name="statusText"
-        type="text"
-        onChange={formik.handleChange}
-        value={formik.values.statusText}
-        placeholder="Обновите свой статус"
-      />
-    </form>
-  );
-};
-
-export default FormStatus;
+export default comp;
